Move route specific types into routes
I got confused when I looked in this file in the RPC folder, and
these are not types for the RPC Layer, but types for individual
routes. They should not live in the RPC code because it's conflating
a system, with things that consume that system.
