+++ ironfish/src/rpc/index.ts
@@ -8,4 +8,4 @@ export * from './response'
 export * from './routes'
 export * from './server'
 export * from './stream'
-export * from './types'
+export * from './routes/types'
