+++ ironfish/src/rpc/routes/wallet/mintAsset.ts
@@ -11,9 +11,9 @@ import { Assert } from '../../../assert'
 import { CurrencyUtils, YupUtils } from '../../../utils'
 import { MintAssetOptions } from '../../../wallet/interfaces/mintAssetOptions'
 import { ValidationError } from '../../adapters'
-import { RpcAsset, RpcAssetSchema, RpcMint, RpcMintSchema } from '../../types'
 import { ApiNamespace } from '../namespaces'
 import { routes } from '../router'
+import { RpcAsset, RpcAssetSchema, RpcMint, RpcMintSchema } from '../types'
 import { RpcWalletTransaction, RpcWalletTransactionSchema } from './types'
 import { getAccount, serializeRpcWalletTransaction } from './utils'
